+++ src/test/java/com/github/jlangch/venice/sandbox/Sandbox_MaxExecTime_Test.java
@@ -71,7 +71,7 @@ public class Sandbox_MaxExecTime_Test {
 		assertThrows(SecurityException.class, () -> venice.eval(script));
 		
 		final long elapsed = sw.stop().elapsedMillis();
-		assertTrue(2000 < elapsed && elapsed < 2500);
+		assertTrue(2000 < elapsed && elapsed < 3000);
 	}
 
 	@Test