+++ ChangeLog.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - added a result history to the REPL `*1`, `*2`, `*3` return the last, the second 
         last and the third last result
+- added REPL help 
 - reworked meta data handling. E.g.: `(def ^:test m [1 2 3])`, `(def ^{:test true} m [1 2 3])`