+++ ChangeLog.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 
 
-## [1.3.6] - 2019-03-xx
+## [1.3.6] - 2019-03-02
 
 ### added